Minor changes to codebase
* Tightened tsconfig 
* Fixded typing issues
* Removed package.json from top level
* Adjusted setup.py
* Adjusted MANIFEST.in
