+++ MANIFEST.in
@@ -1,9 +1,7 @@
 include webviz_core_components/webviz_core_components.min.js
 include webviz_core_components/webviz_core_components.dev.js
 include webviz_core_components/webviz_core_components.css
-include webviz_core_components/plotly-cartesian.js
 include webviz_core_components/metadata.json
 include webviz_core_components/package.json
 include README.md
 include LICENSE
-include react/package.json
