--- package.json
@@ -1,101 +1,0 @@
-{
-    "name": "@webviz/core-components",
-    "description": "Core components for webviz-config",
-    "repository": {
-        "type": "git",
-        "url": "git://github.com/equinor/webviz-core-components.git"
-    },
-    "bugs": {
-        "url": "https://github.com/equinor/webviz-core-components/issues"
-    },
-    "homepage": "https://github.com/equinor/webviz-core-components",
-    "main": "dist/index.js",
-    "types": "dist/index.d.ts",
-    "files": [
-        "dist/**/*"
-    ],
-    "scripts": {
-        "start": "webpack serve --mode development --devtool inline-source-map --entry ./src/demo/index.tsx --open",
-        "transpile": "tsc && npm run copy-files",
-        "transpile:dash": "tsc && rimraf ./dist/components/*/**/",
-        "tidy-up": "rimraf ../R && rimraf ./dist && rimraf ../inst",
-        "build:js": "webpack --mode production --entry ./dist/index.js",
-        "build:js-dev": "webpack --mode development --entry ./dist/index.js",
-        "build:py_and_r": "npm run transpile:dash && cp ./package.json ../webviz_core_components/package.json && cp ./package.json ../package.json && cd ../ && dash-generate-components ./react/dist webviz_core_components -p package.json --r-prefix '' --ignore index.js$",
-        "build": "npm run transpile && npm run build:js && npm run build:js-dev && npm run tidy-up && npm run build:py_and_r && npm run tidy-up",
-        "build:npm": "tsc --declaration && npm run copy-files",
-        "copy-files": "copyfiles -u 2 ./src/lib/**/*.svg ./src/lib/**/*.css ./dist/",
-        "prepare": "npm run build:npm",
-        "postpack": "npm run tidy-up",
-        "prepublishOnly": "npm test && npm run lint",
-        "preversion": "npm run lint",
-        "typecheck": "tsc --noEmit",
-        "format": "eslint --fix 'src/**/*.+(ts|tsx|js|jsx|json|css)' 'tests/**/*.+(ts|tsx|js|jsx|json|css)'",
-        "lint": "eslint 'src/**/*.+(ts|tsx|js|jsx|json|css)' 'tests/**/*.+(ts|tsx|js|jsx|json|css)'",
-        "validate": "npm run typecheck && npm run lint",
-        "test": "jest"
-    },
-    "author": "Equinor <noreply@equinor.com>",
-    "license": "MIT",
-    "dependencies": {
-        "@fortawesome/fontawesome-svg-core": "^1.2.35",
-        "@fortawesome/free-solid-svg-icons": "^5.15.3",
-        "@fortawesome/react-fontawesome": "^0.1.14",
-        "@material-ui/core": "^4.11.3",
-        "@material-ui/lab": "^4.0.0-alpha.57",
-        "classnames": "^2.2.6",
-        "html2canvas": "^1.0.0-rc.7",
-        "notistack": "^1.0.5",
-        "react-colorscales": "^0.7.3",
-        "reactour": "^1.18.3",
-        "styled-components": "^5.2.1"
-    },
-    "devDependencies": {
-        "@babel/core": "^7.13.10",
-        "@babel/plugin-proposal-class-properties": "^7.13.0",
-        "@babel/preset-env": "^7.13.12",
-        "@babel/preset-react": "^7.12.13",
-        "@babel/preset-typescript": "^7.13.0",
-        "@testing-library/react": "^11.2.5",
-        "@testing-library/user-event": "^13.0.8",
-        "@types/classnames": "^2.2.11",
-        "@types/jest": "^26.0.21",
-        "@types/react": "^17.0.3",
-        "@types/react-dom": "^17.0.3",
-        "@types/reactour": "^1.18.1",
-        "@typescript-eslint/eslint-plugin": "^4.19.0",
-        "@typescript-eslint/parser": "^4.19.0",
-        "babel-jest": "^26.6.3",
-        "babel-loader": "^8.2.2",
-        "babel-plugin-transform-es2015-modules-commonjs": "^6.26.2",
-        "copyfiles": "^2.4.1",
-        "css-loader": "^5.1.3",
-        "css-minimizer-webpack-plugin": "^1.3.0",
-        "eslint": "^7.22.0",
-        "eslint-config-prettier": "^8.1.0",
-        "eslint-plugin-import": "^2.22.1",
-        "eslint-plugin-react": "^7.23.0",
-        "file-loader": "^6.2.0",
-        "jest": "^26.6.3",
-        "mini-css-extract-plugin": "^1.3.9",
-        "prop-types": "^15.7.2",
-        "react": "^17.0.2",
-        "react-docgen": "^5.3.1",
-        "react-dom": "^17.0.2",
-        "react-test-renderer": "^17.0.2",
-        "rimraf": "^2.7.1",
-        "style-loader": "^2.0.0",
-        "styled-jsx": "^3.4.4",
-        "ts-loader": "^8.0.18",
-        "tsconfig.js": "^3.0.0",
-        "typescript": "^4.2.3",
-        "url-loader": "^4.1.1",
-        "webpack": "^5.27.2",
-        "webpack-cli": "^4.5.0",
-        "webpack-dev-server": "^3.11.2"
-    },
-    "engines": {
-        "node": ">=12.14.0",
-        "npm": ">=6.14.6"
-    }
-}
