+++ react/package.json
@@ -21,7 +21,7 @@
     "tidy-up": "rimraf ./R && rimraf ./dist && rimraf ./inst && rimraf ./man && rimraf ./.Rbuildignore",
     "build:js": "webpack --mode production --entry ./dist/index.js",
     "build:js-dev": "webpack --mode development --entry ./dist/index.js",
-    "build:py_and_r": "npm run transpile:dash && cp ./package.json ./webviz_core_components/package.json && dash-generate-components ./dist webviz_core_components -p package.json --r-prefix '' --ignore index.js$ && cp -a ./webviz_core_components/. ../webviz_core_components/ && rimraf ./webviz_core_components",
+    "build:py_and_r": "npm run transpile:dash && mkdir -p ./webviz_core_components && cp ./package.json ./webviz_core_components/package.json && dash-generate-components ./dist/components webviz_core_components -p package.json --r-prefix '' --ignore '(^index.js|.stories.js)$' && cp -a ./webviz_core_components/. ../webviz_core_components/ && rimraf ./webviz_core_components",
     "build": "npm run transpile && npm run build:js && npm run build:js-dev && npm run tidy-up && npm run build:py_and_r && npm run tidy-up",
     "build:npm": "tsc --declaration && npm run copy-files",
     "copy-files": "copyfiles -u 2 ./src/lib/**/*.svg ./src/lib/**/*.css ./dist/",
@@ -65,7 +65,6 @@
     "@storybook/react": "^6.2.9",
     "@testing-library/react": "^11.2.5",
     "@testing-library/user-event": "^13.0.8",
-    "@types/classnames": "^2.2.11",
     "@types/jest": "^26.0.23",
     "@types/react": "^17.0.3",
     "@types/react-dom": "^17.0.3",
