+++ react/src/custom_typings/declarations.d.ts
@@ -1,0 +1,3 @@
+// This file contains ambient type declarations for npm modules without type declarations
+
+declare module 'react-colorscales';
