--- react/src/stories/Header.stories.tsx
@@ -1,19 +1,0 @@
-import React from 'react';
-import { Story, Meta } from '@storybook/react';
-
-import { Header, HeaderProps } from './Header';
-
-export default {
-  title: 'Example/Header',
-  component: Header,
-} as Meta;
-
-const Template: Story<HeaderProps> = (args) => <Header {...args} />;
-
-export const LoggedIn = Template.bind({});
-LoggedIn.args = {
-  user: {},
-};
-
-export const LoggedOut = Template.bind({});
-LoggedOut.args = {};
