+++ react/tsconfig.json
@@ -8,17 +8,30 @@
         },
         "outDir": "./dist", // path to output directory
         "sourceMap": true, // allow sourcemap support
-        "strictNullChecks": true, // enable strict null checks as a best practice
+        "strict": true, // enable strict null checks as a best practice
+        // These (and more) are turned on by strict
+        //"strictNullChecks": true, // enable strict null checks as a best practice
+        //"noImplicitAny": true,
+        //"noImplicitThis": true,
+        //"strictPropertyInitialization": true,
+        // Should probable have these 'linting' options enabled
+        "noImplicitReturns": true,
+        "noUnusedLocals": true,
+        "noUnusedParameters": true,
+        "noFallthroughCasesInSwitch": true,
+        "noPropertyAccessFromIndexSignature": true,
+        "forceConsistentCasingInFileNames": true,
         "module": "es6", // specify module code generation
         "moduleResolution": "node",
         "allowSyntheticDefaultImports": true,
         "esModuleInterop": true,
         "jsx": "react", // use typescript to transpile jsx to js
-        "target": "es5", // specify ECMAScript target version
+        "target": "es6", // specify ECMAScript target version
         "allowJs": true, // allow a partial TypeScript and JavaScript codebase
         "skipLibCheck": true, // don't do typechecking on the node modules
     },
     "include": [
-        "./src/lib"
+        "./src/lib",
+        "./src/custom_typings"
     ]
 }
